requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
attrs==21.4.0
Babel==2.10.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
certifi==2022.6.15
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046055389/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.13.0
conda-content-trust @ file:///home/conda/feedstock_root/build_artifacts/conda-content-trust_1621370699668/work
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1649385069444/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967134882/work
cycler==0.11.0
debugpy==1.6.2
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
Django==4.1b1
docutils==0.19b1
entrypoints==0.4
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1646044401614/work
fastjsonschema==2.15.3
fonttools==4.33.3
greenlet==1.1.2
h5py==3.7.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1653754934243/work
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1649067096717/work
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.7.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1654091934149/work/libmambapy
lxml==4.9.1
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1654091934149/work/mamba
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.23.0
obspy==1.3.0
openquake.engine==3.14.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.1.1
plotly @ file:///opt/conda/envs/env/conda-bld/plotly_1656006748598/work/packages/python/plotly/dist/plotly-5.9.0.tar.gz
prometheus-client==0.14.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1656332401605/work
prov==2.0.0
psutil==5.9.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1649384802055/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.18.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857263093/work
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
rdflib==6.1.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1654796491502/work
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1653464393916/work
scipy==1.9.0rc1
Send2Trash==1.8.0
Shapely==1.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smtk===0.9.0-git7948682
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1655315839047/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1626090218611/work
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1655411388954/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
voila==0.3.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
widgetsnbextension==3.6.1
zipp==3.8.0
